46 Prepwork for pool playback
