+++ server/src/api/application.py
@@ -1,17 +1,33 @@
+from contextlib import asynccontextmanager
 from logging import getLogger
 
+from apscheduler.jobstores.memory import MemoryJobStore
+from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
 from api import pool, search, auth
 
-
 _logger = getLogger("main.application")
+
+
+@asynccontextmanager
+async def setup_scheduler(_: FastAPI):
+    _logger.info("Setting up scheduled jobs")
+    job_stores = {
+        "default": MemoryJobStore()
+    }
+    _logger.debug("Creating and starting scheduler")
+    scheduler = AsyncIOScheduler(jobstores=job_stores)
+    scheduler.start()
+    _logger.debug("Adding cleanup state strings job")
+    scheduler.add_job(auth.cleanup_state_strings, "interval", minutes=1)
+    yield
 
 
 def create_app() -> FastAPI:
     _logger.debug("Creating FastAPI application")
-    application = FastAPI(lifespan=auth.setup_scheduler)
+    application = FastAPI(lifespan=setup_scheduler)
 
     _logger.debug("Adding routers")
     application.include_router(auth.router)
